+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.0.0-alpha.81",
-  "contentSha256_12": "0a851f2bb5f9",
+  "pkgVersion": "1.0.0-alpha.82",
+  "contentSha256_12": "ae334922bb49",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.0.0-alpha.81";
-export const contentSha256_12 = "0a851f2bb5f9";
+export const pkgVersion = "1.0.0-alpha.82";
+export const contentSha256_12 = "ae334922bb49";
 export const file = "nodalia-cards.bundle.js";